Revert "Merge branch 'TeamD/Refactor5/KMK002' into TeamZ/UK5/KAFS20"
This reverts commit 41f268308acf71a2ca1f311adbf9492e91b13228, reversing
changes made to 9c76162291701b1dcdbf3d27c3ecddeef01aa1b5.

+++ nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/dailyperformance/correction/dto/DPAttendanceItemControl.java
@@ -3,8 +3,6 @@
  */
 package nts.uk.screen.at.app.dailyperformance.correction.dto;
 
-import java.math.BigDecimal;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -25,7 +23,7 @@ public class DPAttendanceItemControl {
 	// 3:15分
 	// 4:30分
 	// 5:60分
-	private BigDecimal timeInputUnit;
+	private Integer timeInputUnit;
 
 	private String headerBackgroundColor;
